admin update

+++ time-off-management/src/main/java/yu/cs/spring/model/transaction/repo/LeaveApplicationRepo.java
@@ -3,6 +3,6 @@ package yu.cs.spring.model.transaction.repo;
 import yu.cs.spring.model.BaseRepository;
 import yu.cs.spring.model.transaction.entity.LeaveApplication;
 
-public interface LeaveApplicationRepo extends BaseRepository<LeaveApplication, Integer> {
+public interface LeaveApplicationRepo extends BaseRepository<LeaveApplication, Long> {
 
 }
